@@ -1,19 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <TargetFrameworks>net5.0;net6.0;net7.0;net8.0</TargetFrameworks>
-        <PackageId>Treyt.Yarp.ReverseProxy.Swagger</PackageId>
-        <Version>3.9.0</Version>
-        <Authors>Andrei Tsaregorodtsev</Authors>
-        <PackageProjectUrl>https://github.com/andreytreyt/yarp-swagger</PackageProjectUrl>
-    </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>net10.0</TargetFrameworks>
+		<PackageId>Treyt.Yarp.ReverseProxy.Swagger</PackageId>
+		<Version>4.0.0</Version>
+		<Authors>Andrei Tsaregorodtsev</Authors>
+		<PackageProjectUrl>https://github.com/andreytreyt/yarp-swagger</PackageProjectUrl>
+	</PropertyGroup>
 
-    <ItemGroup>
-        <PackageReference Include="IdentityModel.AspNetCore" Version="4.3.0" />
-        <PackageReference Include="Microsoft.OpenApi.Readers" Version="1.6.10" />
-        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.5.0" />
-        <PackageReference Include="Yarp.ReverseProxy" Version="1.1.2" Condition=" '$(TargetFramework)' == 'net5.0' " />
-        <PackageReference Include="Yarp.ReverseProxy" Version="2.1.0" Condition=" '$(TargetFramework)' == 'net6.0' or '$(TargetFramework)' == 'net7.0' or '$(TargetFramework)' == 'net8.0' " />
-    </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="IdentityModel.AspNetCore" Version="4.3.0" />
+		<PackageReference Include="Microsoft.OpenApi.Readers" Version="1.6.28" />
+		<PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="10.1.1" />
+		<PackageReference Include="Yarp.ReverseProxy" Version="2.3.0" />
+	</ItemGroup>
 
 </Project>
